requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-dev --no-hashes --format requirements-txt
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
bcrypt==5.0.0
    # via th-smartsearch
certifi==2026.2.25
    # via
    #   elastic-transport
    #   httpcore
    #   httpx
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
click==8.3.1
    # via
    #   nltk
    #   uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
contourpy==1.3.3
    # via matplotlib
cryptography==46.0.5
    # via python-jose
cycler==0.12.1
    # via matplotlib
datasketch==1.9.0
    # via th-smartsearch
dawg2-python==0.9.0
    # via pymorphy3
ecdsa==0.19.1
    # via python-jose
elastic-transport==8.17.1
    # via elasticsearch
elasticsearch==8.19.3
    # via th-smartsearch
et-xmlfile==2.0.0
    # via openpyxl
faiss-cpu==1.13.2
    # via th-smartsearch
fastapi==0.135.1
    # via th-smartsearch
fonttools==4.62.1
    # via matplotlib
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via th-smartsearch
idna==3.11
    # via
    #   anyio
    #   httpx
joblib==1.5.3
    # via nltk
kiwisolver==1.5.0
    # via matplotlib
lxml==6.0.2
    # via python-docx
matplotlib==3.10.8
    # via th-smartsearch
nltk==3.9.3
    # via th-smartsearch
numpy==2.4.3
    # via
    #   contourpy
    #   datasketch
    #   faiss-cpu
    #   matplotlib
    #   pandas
    #   scipy
    #   th-smartsearch
openpyxl==3.1.5
    # via th-smartsearch
packaging==26.0
    # via
    #   faiss-cpu
    #   matplotlib
pandas==3.0.1
    # via th-smartsearch
pillow==12.1.1
    # via matplotlib
pyasn1==0.6.2
    # via
    #   python-jose
    #   rsa
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via th-smartsearch
pymorphy3==2.0.6
    # via th-smartsearch
pymorphy3-dicts-ru==2.4.417150.4580142
    # via pymorphy3
pyparsing==3.3.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   elasticsearch
    #   matplotlib
    #   pandas
python-docx==1.2.0
    # via th-smartsearch
python-dotenv==1.2.2
    # via pydantic-settings
python-jose==3.5.0
    # via th-smartsearch
rapidfuzz==3.14.3
    # via th-smartsearch
redis==7.3.0
    # via th-smartsearch
regex==2026.2.28
    # via nltk
rsa==4.9.1
    # via python-jose
scipy==1.17.1
    # via datasketch
setuptools==82.0.1
    # via pymorphy3
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
starlette==0.52.1
    # via fastapi
tqdm==4.67.3
    # via nltk
typing-extensions==4.15.0
    # via
    #   elasticsearch
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
tzdata==2025.3 ; sys_platform == 'emscripten' or sys_platform == 'win32'
    # via pandas
urllib3==2.6.3
    # via elastic-transport
uvicorn==0.41.0
    # via th-smartsearch
